requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
yfinance>=0.2.0
gym>=0.26.0
pyyaml>=6.0